requirements.txt
Flask==3.0.0
Pillow==10.1.0
Werkzeug==3.0.1
numpy==1.24.3
